Implement dynamic array clone in value propagation
Dynamic array cloning transforms refArrayObject.clone()
into anewarray followed by arraycopy, where refArrayObject
is a resolved, not fixed array class. This is safe as cloning
of an array object can safely be devirtualized and replaced with
a new array allocation followed by an array copy. The caveat is
that since the array class is not fixed, the actual class type of the
cloned array cannot be known at compile time, hence the class
child of the anewarray opcode needs to be dynamically evaluated
at runtime, instead of being represented as a runtime constant under
normal, statically allocated anewarray cases.

Signed-off-by: Yan Luo <Yan_Luo@ca.ibm.com>