Node API computeIsCollectedReference support for ternary node
Current implementation of Node::computeIsCollectedReference cannot handle
ternary nodes as it is unaware of the two value children, both of which
may need to be checked in order to answer its collected-ness. This change
provides support for ternary nodes by recursively calling the API on the
true-value child, and then, if the answer is inconclusive (i.e. true-value
node is a NULL constant), on the false-value child node. The recursion
continues if a value child is itself a ternary node. The existing
implementation logic is preserved and reused in the tail call to answer
the collected-ness of the value child nodes. The current implementation of
computeIsCollectedReference is refectored into a private method
`computeIsCollectedReferenceImpl` which preserves the logic and existing
behavior for all supported nodes and answers either `TR_yes` or `TR_no`;
in adding support for ternary nodes, a `TR_maybe` answer is possible if
both its value children returns inconclusive answer - e.g. when both
children are NULL constant, which is not impossible during some
optimization pass. Under such case, a `TR_maybe` is considered not
collected.

Signed-off-by: Yan Luo <Yan_Luo@ca.ibm.com>